Revert "fix: fix calculate Value (USDT) and Unrealized PnL (USDT) in hover donut chart"
This reverts commit f027cb066b65fc682e4d158c3313e6d6598d73b5.
